// apps/ui/aiwb/enums/namespace-workloads-table-field.ts
// Copyright © Advanced Micro Devices, Inc., or its affiliates.
//
// SPDX-License-Identifier: MIT

export enum NamespaceWorkloadsTableField {
  NAME = 'displayName',
  TYPE = 'type',
  STATUS = 'status',
  GPUS = 'gpuCount',
  VRAM = 'vram',
  CREATED_AT = 'createdAt',
  CREATED_BY = 'createdBy',
}